# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Contact {
  id: number;
  name: string;
  walletAddress: string;
  /** @nullable */
  ensName?: string | null;
  /** @nullable */
  ownerAddress?: string | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  walletAddress: string;
  ensName?: string;
  ownerAddress?: string;
}

export type PaymentRequestStatus = typeof PaymentRequestStatus[keyof typeof PaymentRequestStatus];


export const PaymentRequestStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface PaymentRequest {
  id: string;
  recipientAddress: string;
  amount: string;
  token: string;
  /** @nullable */
  memo?: string | null;
  status: PaymentRequestStatus;
  /** @nullable */
  paidTxHash?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export type PaymentRequestInputToken = typeof PaymentRequestInputToken[keyof typeof PaymentRequestInputToken];


export const PaymentRequestInputToken = {
  USDC: 'USDC',
  ETH: 'ETH',
} as const;

export interface PaymentRequestInput {
  /** @minLength 1 */
  recipientAddress: string;
  /** @minLength 1 */
  amount: string;
  token: PaymentRequestInputToken;
  memo?: string;
}

export type PaymentRequestUpdateStatus = typeof PaymentRequestUpdateStatus[keyof typeof PaymentRequestUpdateStatus];


export const PaymentRequestUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface PaymentRequestUpdate {
  status?: PaymentRequestUpdateStatus;
  paidTxHash?: string;
}

export interface PaymentStats {
  totalRequestsCreated: number;
  totalRequestsPaid: number;
  totalAmountRequested: string;
  pendingRequests: number;
  recentActivity: PaymentRequest[];
}

export interface DeleteResult {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type ListContactsParams = {
ownerAddress?: string;
};

export type ListPaymentRequestsParams = {
recipientAddress?: string;
};