# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  walletAddress: string;
  /** @nullable */
  ensName?: string | null;
  /** @nullable */
  ownerAddress?: string | null;
  createdAt: string;
}