# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResult {
  success: boolean;
}