# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestStatus } from './paymentRequestStatus';

export interface PaymentRequest {
  id: string;
  recipientAddress: string;
  amount: string;
  token: string;
  /** @nullable */
  memo?: string | null;
  status: PaymentRequestStatus;
  /** @nullable */
  paidTxHash?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}