# lib/api-zod/src/generated/types/paymentRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestInputToken } from './paymentRequestInputToken';

export interface PaymentRequestInput {
  /** @minLength 1 */
  recipientAddress: string;
  /** @minLength 1 */
  amount: string;
  token: PaymentRequestInputToken;
  memo?: string;
}