# lib/api-zod/src/generated/types/paymentRequestUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestUpdateStatus } from './paymentRequestUpdateStatus';

export interface PaymentRequestUpdate {
  status?: PaymentRequestUpdateStatus;
  paidTxHash?: string;
}