# lib/api-zod/src/generated/types/paymentStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BasePay — USDC Payments on Base Network
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequest } from './paymentRequest';

export interface PaymentStats {
  totalRequestsCreated: number;
  totalRequestsPaid: number;
  totalAmountRequested: string;
  pendingRequests: number;
  recentActivity: PaymentRequest[];
}